==> analysis_code/241018_pre_ml/template/01.pre_dmr.ipynb <==
# Parameters
cpu = 8
group_name = "Astro-TE_NN"
mem_gb = 10


import pandas as pd
import subprocess
import numpy as np

#group_name = 'DG_Glut'

home_dir = '/home/qzeng_salk_edu/project/241018_pre_ml'
ct = group_name

dmr_dir = '/data/mC_FILTER_DMR_BED'
_type = 'Total'

dmr_bf = pd.read_csv(f'{dmr_dir}/{ct}.{_type}.aDMR.bed', sep='\t', header=None)
dmr_bf.head()

# gene_df=  pd.read_csv('/data/ref/mouse/genome_anno/mm10_gene.bed', sep='\t', header=None)
# # extend 1 and 2 2000kb
# gene_df[1] = gene_df[1] - 2000
# gene_df[2] = gene_df[2] + 2000
# #gene_df.to_csv('mm10_gene_2kb.bed', sep='\t', header=None, index=None) 
# gene_df.head()

# use bedtools intersect to find the genes that overlap with the DMRs
command = f"bedtools intersect -a {dmr_dir}/{ct}.{_type}.aDMR.bed -b {home_dir}/mm10_gene_2kb.bed -wa -wb > {ct}.aDMR_gene.bed"
subprocess.run(command, shell=True)

dmr_gene_df = pd.read_csv(f'{ct}.aDMR_gene.bed', sep='\t', header=None)
# add a columns that is np.log2([6] - [4])
dmr_gene_df.columns = ['chr','start','end','dmr_id','2mo','9mo','18mo', 
                        'gene_chr','gene_start','gene_end','gene_id','gene_strand','gene_name','gene_type']
dmr_gene_df['old-young'] = dmr_gene_df['18mo'] -dmr_gene_df['2mo']
dmr_gene_df.head()


subprocess.run(f"rm {ct}.aDMR_gene.bed", shell=True)

# drop dmr_id, gen_id, gene_strand, old-young
dmr_gene_df['distance'] = np.minimum(abs(dmr_gene_df['start'] - dmr_gene_df['gene_start']), abs(dmr_gene_df['start'] - dmr_gene_df['gene_end']))
dmr_gene_df = dmr_gene_df.drop(columns=['chr','gene_chr','dmr_id','gene_id','gene_strand','old-young','gene_type'])
dmr_gene_df.head()

dmr_gene_df.to_csv(f'{home_dir}/ml_input/{ct}/{ct}.aDMR_gene.csv', index=None)

## load enhancer

mc_abc_dir = '/data/female-amb/ABC.Enhancer/Re.ABC.By.Methylation'

mouse_genes = pd.read_csv(f"/home/qzeng_salk_edu/project/240321_abc_score/use.mouse_genes.csv", index_col =0)

def load_abc_df(ct, age):
    df = pd.read_csv(f'{mc_abc_dir}/{ct}/filtered.{ct}.{age}.abc_score.csv', index_col =0)
    if age == '8wk':
        age = '2mo'
    df.columns =  [f'{age}.activity',f'{age}.contact',f'{age}.abc_score']
    return df

young_abc_df = load_abc_df(ct, '8wk')
middle_abc_df = load_abc_df(ct, '9mo')
old_abc_df = load_abc_df(ct, '18mo')

# concat the three age groups
abc_df = pd.concat([young_abc_df, middle_abc_df, old_abc_df], axis=1)
#abc_df = abc_df.dropna()#
#abc_df = abc_df.fillna(0)

abc_df['chr'] = abc_df.index.map(lambda x: x.split('-')[0])
abc_df['start'] = abc_df.index.map(lambda x: int(x.split('-')[1]))
abc_df['end'] = abc_df.index.map(lambda x: int(x.split('-')[2]))
abc_df['gene_id'] = abc_df.index.map(lambda x: x.split('-')[3])
abc_df['gene_name'] = abc_df['gene_id'].map(lambda x: mouse_genes.loc[x,'gene_name'])

abc_df['gene_start'] = abc_df['gene_id'].map(lambda x: mouse_genes.loc[x,'start'])
abc_df['gene_end'] = abc_df['gene_id'].map(lambda x: mouse_genes.loc[x,'end'])
#abc_df['gene_length'] = abc_df['gene_end'] - abc_df['gene_start']
#abc_df['contact_distance'] = abs(abc_df['end'] - abc_df['start'])
# add eg_distance, based on the start is more close to gene_start ot gene_end; use the smaller absolute value
#abc_df['eg_distance'] = np.minimum(abs(abc_df['start'] - abc_df['gene_start']), abs(abc_df['start'] - abc_df['gene_end']))
#abc_df drop chr, start, end, gene_start, gene_end
#abc_df = abc_df.drop(columns=['chr','start','end','gene_start','gene_end'])
abc_df


abc_df.to_csv(f'{home_dir}/ml_input/{ct}/{ct}.abc_enhancer.DMR_gene.csv', index=None)
