==> analysis_code/241018_pre_ml/template/02.mcg_gene_body.ipynb <==
# Parameters
cpu = 8
group_name = "Astro-TE_NN"
mem_gb = 10


import pandas as pd
import subprocess
import numpy as np

#group_name = 'DG_Glut'

home_dir = '/home/qzeng_salk_edu/project/241018_pre_ml'
ct = group_name

_dir ='/data/dmg_result'

all_dmg = pd.read_csv(f"{_dir}/Male.CGN.DMG.csv", index_col =0)
all_dmg['celltype'] = all_dmg['celltype'].apply(lambda x: x.replace('/','').replace(' ','_'))

gene_meta = pd.read_csv(f"/data/ref/gencode.vM22.basic.annotation.gene.flat.tsv.gz", sep = '\t')
gene_meta.head()

dmg =all_dmg[all_dmg['celltype'] == ct].copy()
dmg['gene_name'] = dmg.index.map(gene_meta.set_index('gene_id')['gene_name'].to_dict())
# change columns name 8wk to 2mo
dmg.columns = dmg.columns.str.replace('8wk', '2mo')
# add a column DMG, make corrected_pvalue < 0.05 and methylation_change > 0 as -1; corrected_pvalue < 0.05 and methylation_change < 0 as 1; the rest as 0
dmg['DMG'] = np.where((dmg['methylation_change'] > 0) & (dmg['pvalue'] < 0.05), -1, np.where((dmg['methylation_change'] < 0) & (dmg['pvalue'] < 0.05), 1, 0))
#dmg drop celltype, methylation_change, abs_change, trend
dmg = dmg.drop(columns = ['celltype', 'methylation_change', 'abs_change', 'trend'])
dmg.head()

dmg.to_csv(f"{home_dir}/ml_input/{ct}/{ct}.mCG_genebody_gene.csv")
