==> analysis_code/241018_pre_ml/template/04.pre_atac_adar.ipynb <==
# Parameters
cpu = 8
group_name = "Astro-TE_NN"
mem_gb = 10


import pandas as pd
import subprocess
import numpy as np
import pyBigWig
from tqdm import tqdm
import seaborn as sns
import matplotlib.pyplot as plt

#group_name = 'DG_Glut'

home_dir = '/home/qzeng_salk_edu/project/241018_pre_ml'
ct = group_name
dar_dir = '/data/combined_DARs_redo'
abc_dir = '/data/celltype_age_RPM_files'
_type = 'up_peaks'

dar_ct_dict = {'L23_IT_CTX_Glut':'L2-3_IT_CTX_Glut'}
bw_ct_dict = {'L23_IT_CTX_Glut':'L2_3_IT_CTX_Glut'}
if ct in bw_ct_dict.keys():
    dar_ct = dar_ct_dict[ct]
    bw_ct = bw_ct_dict[ct]
else:
    dar_ct = ct
    bw_ct = ct

atac_df  = pd.read_csv(f'{dar_dir}/diff_peaks_{dar_ct}_2vs18.csv', sep=',', index_col=0)

# add a chrom, start, end based on the index 
atac_df['chrom'] = atac_df.index.str.split(':').str[0]
atac_df['start'] = atac_df.index.str.split(':').str[1].str.split('-').str[0].astype(int)
atac_df['end'] = atac_df.index.str.split(':').str[1].str.split('-').str[1].astype(int)
atac_df

up_atac_df = pd.read_csv(f'{dar_dir}/up_peaks_{dar_ct}.bed', sep='\t', header = None)
#add index like in the atac_df
up_atac_df.index = up_atac_df[0] + ':' + up_atac_df[1].astype(str) + '-' + up_atac_df[2].astype(str)
# add the log(fold_change) baes on the info in atac_df
up_atac_df['log2FoldChange'] = atac_df['log2(fold_change)']
up_atac_df.head()

# write the same for down_peaks
down_atac_df = pd.read_csv(f'{dar_dir}/down_peaks_{dar_ct}.bed', sep='\t', header = None)
down_atac_df.index = down_atac_df[0] + ':' + down_atac_df[1].astype(str) + '-' + down_atac_df[2].astype(str)
down_atac_df['log2FoldChange'] = atac_df['log2(fold_change)']
down_atac_df.head()

# combina the down and up
atac_df = pd.concat([up_atac_df, down_atac_df])
# save atac_df to bed
atac_df.to_csv(f'{ct}.aDAR.bed', sep='\t', header=False, index=False)

# use bedtools intersect to find the genes that overlap with the DMRs
command = f"bedtools intersect -a {ct}.aDAR.bed -b {home_dir}/mm10_gene_2kb.bed -wa -wb > {ct}.aDAR_gene.bed"
subprocess.run(command, shell=True)

subprocess.run(f"rm {ct}.aDAR.bed", shell=True)

dar_gene_df = pd.read_csv(f'{ct}.aDAR_gene.bed', sep='\t', header=None)
dar_gene_df.columns = ['chr','start','end','log2(old/young)','gene_chr','gene_start','gene_end','gene_id','gene_strand','gene_name','gene_type']
dar_gene_df.head()

atac_rmp = pd.read_csv(f"{abc_dir}/{ct}_RPM.txt", sep = '\t')
# add three coolumns chr, start and end based on ethe index 
atac_rmp['chr'] = atac_rmp.index.str.split(':').str[0]
atac_rmp['start'] = atac_rmp.index.str.split(':').str[1].str.split('-').str[0].astype(int)
atac_rmp['end'] = atac_rmp.index.str.split(':').str[1].str.split('-').str[1].astype(int)
atac_rmp

dar_gene_df.index = dar_gene_df['chr'] + ':' + dar_gene_df['start'].astype(str) + '-' + dar_gene_df['end'].astype(str)

shared_index = dar_gene_df.index.intersection(atac_rmp.index)
len(shared_index)

dar_gene_df['2mo'] = atac_rmp.iloc[:,0].to_dict()
dar_gene_df['9mo'] = atac_rmp.iloc[:,1].to_dict()
dar_gene_df['18mo'] = atac_rmp.iloc[:,2].to_dict()

non_na_dar_gene_df = dar_gene_df[~dar_gene_df.isna().any(axis=1)]
with_na_dar_gene_df =  dar_gene_df[dar_gene_df.isna().any(axis=1)]

# use PyBigwig to get the atac information for the bw files
bw_dir = "/data/from_luisa/female_atac"
age_list = ['2mo', '9mo', '18mo']
total_counts_df = {}
for age in age_list:
    bw = pyBigWig.open(f"{bw_dir}/Female_{bw_ct}_{age}_merge.bw")
    total_counts = []
    for dmr_id in tqdm(with_na_dar_gene_df.index):
        _chr, start, end = dmr_id.split(':')[0], dmr_id.split(':')[1].split('-')[0], dmr_id.split(':')[1].split('-')[1]
        start, end = int(start), int(end)
        _count = bw.stats(_chr, start, end, type="mean")
        total_counts.append(_count[0])
        
    total_counts_df[age] = total_counts
total_counts_df = pd.DataFrame(total_counts_df)
    

for age in age_list:
    with_na_dar_gene_df[age] = total_counts_df[age].values

with_na_dar_gene_df['log2(18mo/2mo)'] = np.log2(with_na_dar_gene_df['18mo']/with_na_dar_gene_df['2mo'])

# map rows in dar_gene_df that is in with_na_dar_gene_df.index, give 2mo column values based on with_na_dar_gene_df['2mo'].to_dict()
dar_gene_df.loc[with_na_dar_gene_df.index, '2mo'] = with_na_dar_gene_df['2mo'].to_dict()
dar_gene_df.loc[with_na_dar_gene_df.index, '9mo'] = with_na_dar_gene_df['9mo'].to_dict()
dar_gene_df.loc[with_na_dar_gene_df.index, '18mo'] = with_na_dar_gene_df['18mo'].to_dict()

dar_gene_df.to_csv(f'{home_dir}/ml_input/{ct}/{ct}.aDAR_gene.csv',  index=False)

subprocess.run(f"rm {ct}.aDAR_gene.bed", shell=True)
