==> analysis_code/241018_pre_ml/template/05.pre_atac_rpm.ipynb <==
# Parameters
cpu = 8
group_name = "Astro-TE_NN"
mem_gb = 10


import pandas as pd
import subprocess
import numpy as np

#group_name = 'DG_Glut'

home_dir = '/home/qzeng_salk_edu/project/241018_pre_ml'
ct = group_name
dar_dir = '/data/combined_DARs_redo'
abc_dir = '/data/celltype_age_RPM_files'
#_type = 'up_peaks'

atac_rmp = pd.read_csv(f"{abc_dir}/{ct}_RPM.txt", sep = '\t')
# add three coolumns chr, start and end based on ethe index 
atac_rmp['chr'] = atac_rmp.index.str.split(':').str[0]
atac_rmp['start'] = atac_rmp.index.str.split(':').str[1].str.split('-').str[0].astype(int)
atac_rmp['end'] = atac_rmp.index.str.split(':').str[1].str.split('-').str[1].astype(int)
# put the chr start and end in the first column
cols = atac_rmp.columns.tolist()
cols = cols[-3:] + cols[:-3]
atac_rmp = atac_rmp[cols]
# remove chrY, chrM and chrX
atac_rmp = atac_rmp[~atac_rmp['chr'].isin(['chrY', 'chrM', 'chrX'])]
atac_rmp


# add rows in dar but not atac-rmp
dar_df = pd.read_csv(f'{home_dir}/ml_input/{ct}/{ct}.aDAR_gene.csv')
# make the index as chr:start-end
dar_df.index = dar_df['chr'] + ':' + dar_df['start'].astype(str) + '-' + dar_df['end'].astype(str)
dar_df['log2(18mo/2mo)'] = np.log2((dar_df['18mo'] + 1)/(dar_df['2mo'] + 1))

not_peak_dar =dar_df.loc[~dar_df.index.isin(atac_rmp.index)]
not_peak_dar = not_peak_dar[['chr','start','end','2mo','9mo','18mo']]
not_peak_dar.columns = atac_rmp.columns
atac_rmp = pd.concat([atac_rmp, not_peak_dar], axis = 0)
atac_rmp

atac_rmp.to_csv(f'{ct}.peak.bed', sep = '\t', index = False, header = False)

# use bedtools intersect to find the genes that overlap with the DMRs
command = f"bedtools intersect -a {ct}.peak.bed -b {home_dir}/mm10_gene_2kb.bed -wa -wb > {ct}.ATAC_peak_gene.bed"
subprocess.run(command, shell=True)

peak_gene_df = pd.read_csv(f'{ct}.ATAC_peak_gene.bed', sep='\t', header=None)
peak_gene_df.columns = ['chrom','peak_start','peak_end','2mo','9mo','18mo','gene_chrom','gene_start','gene_end','gene_id','strand', 'gene_name','gene_type']
peak_gene_df.index = peak_gene_df['chrom'] + ':' + peak_gene_df['peak_start'].astype(str) + '-' + peak_gene_df['peak_end'].astype(str)
# if the row index in dar and values in 
peak_gene_df

peak_gene_df['DAR'] = peak_gene_df.index.map(dar_df['log2(18mo/2mo)'].to_dict())
# for values in DAR, if na == 0, if > 0 return 1, if < 0 return -1
peak_gene_df['DAR'] = peak_gene_df['DAR'].fillna(0)
peak_gene_df['DAR'] = np.sign(peak_gene_df['DAR'])
peak_gene_df.head()

# drop dmr_id, gen_id, gene_strand, old-young
peak_gene_df['distance'] = np.minimum(abs(peak_gene_df['peak_start'] - peak_gene_df['gene_start']), abs(peak_gene_df['peak_start'] - peak_gene_df['gene_end']))
peak_gene_df = peak_gene_df.drop(columns=['chrom','peak_start','peak_end','gene_chrom','gene_id','strand','gene_type'])
peak_gene_df.head()

peak_gene_df.to_csv(f"{home_dir}/ml_input/{ct}/{ct}.peak_gene.csv", index = False)

subprocess.run(f"rm {ct}.peak.bed", shell=True)
subprocess.run(f"rm {ct}.ATAC_peak_gene.bed", shell=True)

## enhancer

atac_abc_dir = '/data/female-amb/ABC.Enhancer/ABC.By.ATAC.RPM.Raw.Cool'

mouse_genes = pd.read_csv(f"/home/qzeng_salk_edu/project/240321_abc_score/use.mouse_genes.csv", index_col =0)

def load_abc_df(ct, age):
    df = pd.read_csv(f'{atac_abc_dir}/{ct}/filtered.{ct}.{age}.abc_score.csv', index_col =0)
    if age == '8wk':
        age = '2mo'
    df.columns =  [f'{age}.activity',f'{age}.contact',f'{age}.abc_score']
    return df

young_abc_df = load_abc_df(ct, '8wk')
middle_abc_df = load_abc_df(ct, '9mo')
old_abc_df = load_abc_df(ct, '18mo')

# concat the three age groups
abc_df = pd.concat([young_abc_df, middle_abc_df, old_abc_df], axis=1)
#abc_df = abc_df.dropna()#
#abc_df = abc_df.fillna(0)

abc_df['chr'] = abc_df.index.map(lambda x: x.split(':')[0])
abc_df['start'] = abc_df.index.map(lambda x: int(x.split(':')[1].split('-')[0]))
abc_df['end'] = abc_df.index.map(lambda x: int(x.split(':')[1].split('-')[1]))
abc_df['gene_id'] = abc_df.index.map(lambda x: x.split(':')[1].split('-')[2])
abc_df['gene_name'] = abc_df['gene_id'].map(lambda x: mouse_genes.loc[x,'gene_name'])

abc_df['gene_start'] = abc_df['gene_id'].map(lambda x: mouse_genes.loc[x,'start'])
abc_df['gene_end'] = abc_df['gene_id'].map(lambda x: mouse_genes.loc[x,'end'])
# add eg_distance, based on the start is more close to gene_start ot gene_end; use the smaller absolute value
#abc_df drop chr, start, end, gene_start, gene_end
#abc_df = abc_df.drop(columns=['chr','start','end','gene_start','gene_end'])
abc_df


abc_df.to_csv(f'{home_dir}/ml_input/{ct}/{ct}.abc_enhancer.peak_gene.csv', index=None)
