==> analysis_code/241018_pre_ml/template/06.pre_hic_loop.ipynb <==
# Parameters
cpu = 8
group_name = "Astro-TE_NN"
mem_gb = 10


import pandas as pd
from scipy.stats import norm, zscore
import numpy as np
import seaborn as sns
import subprocess
import matplotlib.pyplot as plt 

#group_name = 'DG_Glut'

home_dir = '/home/qzeng_salk_edu/project/241018_pre_ml'
ct = group_name

loop_dir = '/data/female-amb/Diff.Result/loop_ct_age/diff'

loopq = pd.read_hdf(f'{loop_dir}/{ct}/loop_Q.hdf', key='data')
loopt = pd.read_hdf(f'{loop_dir}/{ct}/loop_T.hdf', key='data')

loopq.columns = [f"{ct}.2mo.Q", f"{ct}.9mo.Q", f"{ct}.18mo.Q"]
loopt.columns = [f"{ct}.2mo.T", f"{ct}.9mo.T", f"{ct}.18mo.T"]

loopall = pd.read_hdf(f"{loop_dir}/{ct}/merged_loop.hdf")
loopall = pd.concat([loopall, loopq, loopt], axis = 1)
loopall.head()

sell = ((zscore(np.log(loopall['Qanova']))>norm.isf(0.2)) & 
        (zscore(np.log(loopall['Tanova']))>norm.isf(0.2)) )
sell.sum()

loopall['Diff_Loop'] = 0
loopall.loc[sell, 'Diff_Loop'] = 1
loopall.to_csv(f"{ct}.anchor1_loop.bed", sep = '\t', header = False, index = False)

# move columns 3 4 5 in to the first three
cols = loopall.columns.tolist()
cols = cols[3:6] + cols[:3] + cols[6:] 
loopall = loopall[cols]
loopall.to_csv(f"{ct}.anchor2_loop.bed", sep = '\t', header = False, index = False)

# so diff loop as two anchors, so two chro:start-end region, i want to get the intersection with another bed file as long as it intersect with eithe ranhocr
command = f"bedtools intersect -a {ct}.anchor1_loop.bed -b {home_dir}/mm10_gene_2kb.bed -wa -wb > {ct}.anchor1_loop_gene.bed"
subprocess.run(command, shell=True)

command = f"bedtools intersect -a {ct}.anchor2_loop.bed -b {home_dir}/mm10_gene_2kb.bed -wa -wb > {ct}.anchor2_loop_gene.bed"
subprocess.run(command, shell=True)

subprocess.run(f"rm {ct}.anchor1_loop.bed", shell=True)
subprocess.run(f"rm {ct}.anchor2_loop.bed", shell=True)

# load the output files and concat
anchor1 = pd.read_csv(f"{ct}.anchor1_loop_gene.bed", sep = '\t', header = None)
anchor2 = pd.read_csv(f"{ct}.anchor2_loop_gene.bed", sep = '\t', header = None)
# concat
anchor = pd.concat([anchor1, anchor2], axis = 0)

anchor.columns = ['chrom','anchor1_start','anchor1_end','chrom','anchor2_start','anchor2_end',
                'Qanova','Eanova','Tanova','2mo.Q','9mo.Q','18mo.Q','2mo.T','9mo.T','18mo.T','Diff_Loop',
                'gene_chrom','gene_start','gene_end','gene_id','strand','gene_name','gene_type']
anchor

subprocess.run(f"rm {ct}.anchor1_loop_gene.bed", shell=True)
subprocess.run(f"rm {ct}.anchor2_loop_gene.bed", shell=True)

# add a distance columns, anchor1_start to gene_start and gene end, use which ever is closer
anchor['anchor1_distance'] = np.minimum(abs(anchor['anchor1_start'] - anchor['gene_start']), abs(anchor['anchor1_start'] - anchor['gene_end']))
anchor['anchor2_distance'] = np.minimum(abs(anchor['anchor2_start'] - anchor['gene_start']), abs(anchor['anchor2_start'] - anchor['gene_end']))
anchor['gene_length'] = anchor['gene_end'] - anchor['gene_start']
# drop chrom	anchor1_start	anchor1_end	chrom	anchor2_start	anchor2_end	, 'gene_chrom', 'gene_start','gene_end', 'strand', 'gene_type'
anchor = anchor.drop(columns = ['chrom','anchor1_start','anchor1_end','chrom','anchor2_start','anchor2_end','gene_chrom', 'gene_start','gene_end', 'strand', 'gene_type'])
anchor

anchor.to_csv(f"{home_dir}/ml_input/{ct}/{ct}.Loop_gene.csv",  header = True, index = False)

import subprocess
#ct = 'L6_IT_CTX_Glut'
#subprocess.run(f"tar -cjvf ml_input/{ct}/{ct}.Loop_gene.csv.gz ml_input/{ct}/{ct}.Loop_gene.csv", shell=True)
subprocess.run(f"gzip -k {home_dir}/ml_input/{ct}/{ct}.Loop_gene.csv", shell=True)

subprocess.run(f"rm {home_dir}/ml_input/{ct}/{ct}.Loop_gene.csv", shell=True)

#anchor.to_csv(f"ml_input/{ct}.diff_loop_gene.csv",  header = True, index = False)
